tangledown/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
sympy==0.7.1
django-wakawaka==0.3
Markdown==2.1.0
